backend/requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# SurrealDB
surrealdb>=1.0.0

# Embedding model
torch>=2.4.0
sentence-transformers>=3.1.0
transformers>=4.48.0

# Document parsing
pypdf>=4.0.0
python-docx>=1.1.0

# Settings
pydantic-settings>=2.0.0
